Frontend refactoring
