+++ frontend/src/views/internal/processes.vue
@@ -1,7 +1,7 @@
 <template>
   <div class="container">
-    <ProcessCards :processes="processes" :year="year" :sign="sign"/>
-    <br>
+    <ProcessCards :processes="processes" :year="year" :sign="sign" />
+    <br />
     <button type="button" @click="newProcess()" class="btn btn-primary">
       Neuer Vorgang
     </button>
@@ -9,14 +9,17 @@
 </template>
 
 <script>
-import ProcessCards from "../processCard.vue";
+import ProcessCards from "../../components/processCard.vue";
 export default {
   name: "Vorgänge",
   data() {
     return {
-      processes: [{type:'A',status:'2',name:"Müller"},{type:'B',status:'3',name:"Ano"}],
+      processes: [
+        { type: "A", status: "2", name: "Müller" },
+        { type: "B", status: "3", name: "Ano" },
+      ],
       year: 2019,
-      sign: "Meier"
+      sign: "Meier",
     };
   },
   components: {
